:truck: Move tokenizers to seperate files and add AbstractTokenizer
