+++ conf/dataset/representation/tokenizer/reaction_tokenizer.yaml
@@ -1,0 +1,1 @@
+_target_: chemtorch.components.preprocessing.tokenizer.reaction_tokenizer.ReactionTokenizer
