--- conf/dataset/representation/tokenizer/simple_tokenizer.yaml
@@ -1,1 +1,0 @@
-_target_: chemtorch.components.preprocessing.tokenizer.simple_tokenizer.SimpleTokenizer
