+++ src/chemtorch/components/preprocessing/tokenizer/tokenizer_defaults.py
@@ -1,0 +1,4 @@
+DEFAULT_MOLECULE_PATTERN = r"(\%\([0-9]{3}\)|\[[^\]]+]|Br?|Cl?|N|O|S|P|F|I|b|c|n|o|s|p|\||\(|\)|\.|=|#|-|\+|\\|\/|:|~|@|\?|>>?|\*|\$|\%[0-9]{2}|[0-9])"
+DEFAULT_UNK_TOKEN = "<UNK>"
+REACTION_SEPARATOR_TOKEN = ">>"
+MOLECULE_SEPARATOR_TOKEN = "."
